requirement.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
Pillow==10.0.1
plyer==2.1.0
psutil==5.9.5
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
